Split MVC into three seperate files
MyController.py, MyModel.py, MyView.py
